Refactor host-functions code.
